remaining $this->get replased with Indi::get()

+++ library/Indi/Controller/Front.php
@@ -454,10 +454,10 @@ class Indi_Controller_Front extends Indi_Controller {
                 $finalORDER = $this->finalORDER($finalWHERE, Indi::get()->sort);
 
                 // Get the rowset, fetched using WHERE and ORDER clauses, and with built LIMIT clause,
-                // constructed with usage of $this->get('limit') and $this->get('page') params
+                // constructed with usage of Indi::get('limit') and Indi::get('page') params
                 $this->rowset = Indi::trail()->model->{
                 'fetch'. (Indi::trail()->model->treeColumn() ? 'Tree' : 'All')
-                }($finalWHERE, $finalORDER, Indi::get('limit'), Indi::get('page'));
+                }($finalWHERE, $finalORDER, (int) Indi::get('limit'), (int) Indi::get('page'));
             }
         }
     }
